# Optimize `bool_or` using `max_boolean`