--- a/arrow/benches/aggregate_kernels.rs
+++ b/arrow/benches/aggregate_kernels.rs
@@ -82,11 +82,17 @@ fn add_benchmark(c: &mut Criterion) {
             .bench_function("max nonnull mixed", |b| {
                 b.iter(|| max_boolean(&nonnull_bools_mixed))
             })
+            .bench_function("or nonnull mixed", |b| {
+                b.iter(|| bool_or(&nonnull_bools_mixed))
+            })
             .bench_function("min nonnull false", |b| {
                 b.iter(|| min_boolean(&nonnull_bools_all_false))
             })
             .bench_function("max nonnull false", |b| {
                 b.iter(|| max_boolean(&nonnull_bools_all_false))
+            })
+            .bench_function("or nonnull false", |b| {
+                b.iter(|| bool_or(&nonnull_bools_all_false))
             })
             .bench_function("min nonnull true", |b| {
                 b.iter(|| min_boolean(&nonnull_bools_all_true))
@@ -94,11 +100,17 @@ fn add_benchmark(c: &mut Criterion) {
             .bench_function("max nonnull true", |b| {
                 b.iter(|| max_boolean(&nonnull_bools_all_true))
             })
+            .bench_function("or nonnull true", |b| {
+                b.iter(|| bool_or(&nonnull_bools_all_true))
+            })
             .bench_function("min nullable mixed", |b| {
                 b.iter(|| min_boolean(&nullable_bool_mixed))
             })
             .bench_function("max nullable mixed", |b| {
                 b.iter(|| max_boolean(&nullable_bool_mixed))
+            })
+            .bench_function("or nullable mixed", |b| {
+                b.iter(|| bool_or(&nullable_bool_mixed))
             })
             .bench_function("min nullable false", |b| {
                 b.iter(|| min_boolean(&nullable_bool_all_false))
@@ -106,11 +118,17 @@ fn add_benchmark(c: &mut Criterion) {
             .bench_function("max nullable false", |b| {
                 b.iter(|| max_boolean(&nullable_bool_all_false))
             })
+            .bench_function("or nullable false", |b| {
+                b.iter(|| bool_or(&nullable_bool_all_false))
+            })
             .bench_function("min nullable true", |b| {
                 b.iter(|| min_boolean(&nullable_bool_all_true))
             })
             .bench_function("max nullable true", |b| {
                 b.iter(|| max_boolean(&nullable_bool_all_true))
+            })
+            .bench_function("or nullable true", |b| {
+                b.iter(|| bool_or(&nullable_bool_all_true))
             });
     }
 }